create modular gateway. Add async requests. Add cpu pool
