+++ infra/k8s/auth-depl.yaml
@@ -3,7 +3,7 @@ kind: Deployment
 metadata:
   name: auth-depl
 spec:
-  replicas: 2
+  replicas: 1
   selector: { matchLabels: { app: auth } }
   template:
     metadata: { labels: { app: auth } }
