+++ infra/k8s/gateway-depl.yaml
@@ -49,7 +49,7 @@ spec:
             failureThreshold: 60 
           resources:
             requests: { cpu: "100m", memory: "256Mi" }
-            limits:   { cpu: "600m", memory: "512Mi" }
+            limits:   { cpu: "1000m", memory: "768Mi" }
 
 ---
 apiVersion: v1
